layers-store - Update for grid classes. - Update intersect Test.
git-svn-id: http://alageospatialportal.googlecode.com/svn/trunk/layers-store@1822 73def43e-0e3b-11df-a5af-b38b39d06e06

+++ src/main/java/org/ala/layers/intersect/SamplingThread.java
@@ -135,7 +135,7 @@ public class SamplingThread extends Thread {
                     if (i > 0) {
                         sb.append("\n");
                     }
-                    gc = classes.get(values[i]);
+                    gc = classes.get((int)values[i]);
                     if (gc != null) {
                         sb.append(gc.getName());
                     } else {
